Move client utils to program
