+++ program/Cargo.toml
@@ -6,7 +6,8 @@ license = "WTFPL"
 publish = false
 
 [dependencies]
-solana-program = "=1.10.0"
+solana-program = "=1.9.4"
+solana-client = "=1.9.4"
 borsh = "0.9.3"
 bytemuck = {version = "1.7.2", features = ["derive"]}
 spl-token = {version = "3.2.0", features = ["no-entrypoint"]}
@@ -20,10 +21,10 @@ test-bpf = []
 
 [dev-dependencies]
 assert_matches = "1.4.0"
-solana-program-test = "=1.10.0"
-solana-sdk = "=1.10.0"
-solana-validator = "=1.10.0"
-solana-logger = "=1.10.0"
+solana-program-test = "=1.9.4"
+solana-sdk = "=1.9.4"
+solana-validator = "=1.9.4"
+solana-logger = "=1.9.4"
 
 [lib]
 crate-type = ["cdylib", "lib"]
