+++ program/src/lib.rs
@@ -1,3 +1,4 @@
+pub mod client;
 pub mod entrypoint;
 pub mod error;
 pub mod instruction;
