+++ rust-client/Cargo.toml
@@ -7,8 +7,8 @@ edition = "2021"
 
 [dependencies]
 buoyant = { path = "../program" }
-solana-sdk = "1.10.0"
-solana-client = "1.10.0"
-solana-program = "1.7.14"
+solana-sdk = "1.9.4"
+solana-client = "1.9.4"
+solana-program = "1.9.4"
 spl-token = {version = "3.2.0", features = ["no-entrypoint"]}
 spl-associated-token-account = {version = "1.0.3", features = ["no-entrypoint"]}
